requirements.txt
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
typing_extensions==4.3.0
ecdsa==0.18.0
pycryptodome==3.17
scapy==2.5.0
termcolor==1.1.0
dnspython==2.4.2